# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Mahak Volunteer Management Platform API
  description: |
    وب‌سرویس REST سامانه مدیریت داوطلبان محک (`mahak-volunteer-api`).

    این مشخصات برای پیاده‌سازی کلاینت توسط تیم دیگر است. منبع مسیرها با
    `GET /api/v1/` و `backend/internal/adapter/http/catalog.go` هم‌تراز است.

    **نقش‌ها**
    - `volunteer` داوطلب
    - `operator` بهره‌بردار / واحد پشتیبانی — همان دسترسی `/admin/*`
    - `admin` ادمین سامانه — همان دسترسی `/admin/*`

    **احراز**
    - JWT: `Authorization: Bearer <token>`
    - یکپارچگی داخلی: `X-Internal-Token: <INTERNAL_API_TOKEN>`
    - وب‌هوک ماموریت: توکن ماموریت در بدنه (`token`) یا هدر `X-Mission-Token`؛ با توکن داخلی قاطی نشود

    خطاها همیشه `{"error":"<پیام فارسی یا انگلیسی>"}` هستند.
  version: "1.1.0"
  contact:
    name: Mahak Volunteer Platform
servers:
  - url: http://localhost:8080
    description: Local
  - url: https://{host}
    description: Deployed
    variables:
      host:
        default: volunteers.example.ir
...
security: []
paths:
...
  /api/v1/auth/otp/send:
    post:
      summary: Send OTP
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OTPSendRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/OTPSendResult'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/auth/otp/verify:
    post:
      summary: Verify OTP
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OTPVerifyRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/AuthResult'
        "201": { description: New user }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/auth/login:
    post:
      summary: Login
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/AuthResult'
        "401": { description: Unauthorized }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/auth/register:
    post:
      summary: Register volunteer (role ignored)
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/AuthResult'
        "409": { description: Email exists }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/auth/external:
    post:
      summary: Map Mahak Auth user
      tags: [Auth]
      security: [{ InternalToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExternalAuthRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/AuthResult'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/me:
    get:
      summary: Current session
      tags: [Session]
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/MeResponse'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/notifications:
    get:
      summary: Notifications
      tags: [Session]
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/NotificationList'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/notifications/{id}/read:
    post:
      summary: Mark one read
      tags: [Session]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/StatusOK'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/notifications/read-all:
    post:
      summary: Mark all read
      tags: [Session]
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/StatusOK'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/tickets/me:
    get:
      summary: My tickets
      tags: [Tickets]
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/TicketList'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/tickets:
    post:
      summary: Create ticket
      tags: [Tickets]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketCreate'
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/Ticket'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/tickets/{id}:
    get:
      summary: Ticket thread
      tags: [Tickets]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/Ticket'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/tickets/{id}/messages:
    post:
      summary: Reply to ticket
      tags: [Tickets]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketReply'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/Ticket'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/skills:
    get:
      summary: Public skill catalog
      tags: [Volunteer]
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/SkillGroupList'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/volunteers/me:
    get:
      summary: My profile
      tags: [Volunteer]
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/Volunteer'
        "400": { $ref: "#/components/responses/Error" }
    put:
      summary: Update profile
      tags: [Volunteer]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileInput'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/Volunteer'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/volunteers/me/submit:
    post:
      summary: Submit for review
      tags: [Volunteer]
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/Volunteer'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/volunteers/me/availability:
    get:
      summary: Availability slots
      tags: [Volunteer]
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/AvailabilityList'
        "400": { $ref: "#/components/responses/Error" }
    put:
      summary: Replace availability
      tags: [Volunteer]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvailabilityInput'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/StatusOK'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/volunteers/me/documents:
    get:
      summary: My documents
      tags: [Volunteer]
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/DocumentList'
        "400": { $ref: "#/components/responses/Error" }
    post:
      summary: Upload document
      tags: [Volunteer]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DocumentUpload'
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/Document'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/volunteers/me/documents/{id}:
    delete:
      summary: Delete document
      tags: [Volunteer]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/StatusOK'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/volunteers/me/skill-proposals:
    get:
      summary: My skill proposals
      tags: [Volunteer]
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/SkillProposalList'
        "400": { $ref: "#/components/responses/Error" }
    post:
      summary: Propose skill
      tags: [Volunteer]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillPropose'
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/SkillProposal'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/tasks:
    get:
      summary: Eligible tasks
      tags: [Tasks]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: skill, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/TaskPage'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/tasks/{id}:
    get:
      summary: Task detail
      tags: [Tasks]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/Task'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/tasks/{id}/accept:
    post:
      summary: Apply / reserve
      tags: [Tasks]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/Assignment'
        "409": { description: Already assigned }
        "422": { description: Not eligible or full }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/assignments/me:
    get:
      summary: My assignments
      tags: [Tasks]
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/AssignmentList'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/assignments/{id}/start:
    post:
      summary: Start remote work
      tags: [Tasks]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/Assignment'
        "400": { $ref: "#/components/responses/Error" }
...
  /api/v1/assignments/{id}/rate:
    post:
      summary: Rate organization
      tags: [Tasks]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VolunteerRating'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/Assignment'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/assignments/{id}/cancel:
    post:
      summary: Cancel my request
      tags: [Tasks]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/Assignment'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/missions:
    get:
      summary: Active missions
      tags: [Missions]
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/MissionList'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/missions/{id}/start:
    post:
      summary: Start mission
      tags: [Missions]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/MissionProgress'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/missions/{id}/progress:
    post:
      summary: Verify mission completion
      tags: [Missions]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/MissionProgress'
        "422": { description: Not verified }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/missions/me:
    get:
      summary: My mission progress
      tags: [Missions]
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/MissionProgressList'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/certificates/me:
    get:
      summary: My certificates
      tags: [Certificates]
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/CertificateList'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/certificates/requests:
    get:
      summary: My certificate requests
      tags: [Certificates]
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/CertificateRequestList'
        "400": { $ref: "#/components/responses/Error" }
    post:
      summary: Create certificate request
      tags: [Certificates]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertificateRequestInput'
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/CertificateRequest'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/certificates/verify/{code}:
    get:
      summary: Public verify
      tags: [Certificates]
      parameters:
        - { name: code, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/Certificate'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/certificates/{code}/pdf:
    get:
      summary: Public PDF
      tags: [Certificates]
      parameters:
        - { name: code, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/admin/dashboard:
    get:
      summary: Staff dashboard
      tags: [StaffVolunteers]
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/Dashboard'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/admin/volunteers:
    get:
      summary: List volunteers
      tags: [StaffVolunteers]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: skill, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string } }
        - { name: attention, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer } }
        - { name: offset, in: query, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/VolunteerPage'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/admin/volunteers/{id}:
    get:
      summary: Volunteer dossier
      tags: [StaffVolunteers]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/VolunteerDossier'
        "400": { $ref: "#/components/responses/Error" }
    put:
      summary: Edit profile
      tags: [StaffVolunteers]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileInput'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/Volunteer'
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/admin/volunteers/{id}/review:
    post:
      summary: Review membership
      tags: [StaffVolunteers]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewInput'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/Volunteer'
        "400": { $ref: "#/components/responses/Error" }
...
  /api/v1/admin/volunteers/{id}/documents:
    get:
      summary: Documents
      tags: [StaffVolunteers]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '$#/components/schemas/DocumentList'
        "400": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    InternalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
...
  schemas:
...
    OTPSendRequest:
      type: object
      required: [phone]
      properties:
        phone: { type: string, example: "09121234567" }
    OTPSendResult:
      type: object
      properties:
        phone: { type: string }
        ttl_seconds: { type: integer }
        resend_after: { type: integer }
        is_new: { type: boolean }
        dev_code: { type: string, description: "Only when OTP_REVEAL=true" }
...